pyproject.toml
[tool.poetry]
name = "rinna-slackbot"
version = "0.1.0"
description = ""
authors = ["Koki Takahashi <hakatasiloving@gmail.com>"]
license = "Apache Licence 2.0"
readme = "README.md"
packages = [{ include = "**/*" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
slack-sdk = "^3.36.0"
regex = "^2025.9.1"
python-dotenv = "^1.1.1"
google-api-python-client = "^2.181.0"
firebase-admin = "^7.1.0"
pystray = "^0.19.4"
azure-cognitiveservices-vision-contentmoderator = "^1.0.0"
google-cloud-pubsub = "^2.31.1"
google-cloud-language = "^2.17.2"
datasets = {extras = ["audio"], version = "^3.6.0"}
evaluate = "^0.4.5"
transformers = "^4.56.0"
scikit-learn = "^1.7.1"
psutil = "^7.0.0"
pytest = "^8.4.2"
tokenizers = "^0.22.0"
pytest-mock = "^3.14.1"
torch = {version = "^2.6.0", source = "torch_cu124"}
gstop = "^0.2.3"
accelerate = "^1.10.1"
pproxy = "^2.7.0"
requests = {extras = ["socks"], version = "^2.32.0"}

# $env:CMAKE_ARGS='-DGGML_BLAS=ON -DGGML_BLAS_VENDOR=OpenBLAS -DBLAS_ROOT="C:\Users\hakatashi\AppData\Local\OpenBLAS-0.3.30-x64" -DBLAS_INCLUDE_DIRS="C:\Users\hakatashi\AppData\Local\OpenBLAS-0.3.30-x64\include"'
# $env:PATH += ";C:\Users\hakatashi\AppData\Local\OpenBLAS-0.3.30-x64\bin"
llama-cpp-python = "^0.3.16"


[[tool.poetry.source]]
name = "torch_cu124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[[tool.poetry.source]]
name = "llama_cpp_cu124"
url = "https://abetlen.github.io/llama-cpp-python/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
